+++ src/helpers/hasFilesToDownload.js
@@ -1,7 +1,8 @@
 function hasFilesToDownload(node) {
   const hasImages = node.images && node.images.length;
   const hasImage = node.image && node.image.length;
-  return hasImages || hasImage ? true : false;
+  const hasProductImages = node.product && hasFilesToDownload(node.product);
+  return hasImages || hasImage || hasProductImages ? true : false;
 }
 
 module.exports = hasFilesToDownload;