+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
-- None
+- Fix an issue where the images of a `Sku`'s expanded product field are not downloaded if the `Sku` has no image itself.
 
 ## [3.0.2](https://github.com/njosefbeck/gatsby-source-stripe/compare/v3.0.1...v3.0.2) - 2019-07-26
 - Remove `mediaType` from StripeObject Gatsby node, to solve bug related to using this plugin in the same project as `gatsby-tranformer-json`. Update relevant test.